pyproject.toml
[tool.poetry]
name = "beat-kokoa"
version = "0.1.0"
description = ""
authors = ["pzpz <idleversecat@gmail.com>"]
readme = "README.md"
packages = [{include = "beat_kokoa"}]

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
